requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
azure-storage-blob==12.19.0
mongoengine==0.27.0
pymongo==4.5.0
flask-mongoengine==1.0.0
Werkzeug==2.3.7
redis==5.0.1
Flask-Caching==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
email-validator==2.0.0
